native delete method in laravel

+++ app/Http/Controllers/PostController.php
@@ -47,9 +47,11 @@ class PostController extends Controller
     public function destroy(Post $post){
         // dd($post);
 
-        if($post->ownedBy(auth()->user())){
-            dd('no');
-        }
+        // if($post->ownedBy(auth()->user())){
+        //     dd('no');
+        // }
+
+        $this->authorize('delete', $post);
         $post->delete();
 
         return back();
